requirements.txt
pyyaml
matplotlib
numpy
pyqtgraph